requirements.txt
sentence-transformers
langchain
django
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain-community
langchain-openai
langchain-experimental
-e .